+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Fernschach-Verwaltung Changelog
 
+## Version 0.20.3 (2024-07-08)
+
+* Change: Model HauptkontoModel zu Hauptkonto
+* Add: Model Spieler für den Zugriff auf tl_fernschach_spieler
+
 ## Version 0.20.2 (2024-07-07)
 
 * Add: Klasse ResetUtil für Kontoresets -> Optimierung des Codes, weil diverse alte und neue Funktionen herumgeistern. Suffix Util, weil Reset ein reserviertes Wort in PHP ist
 * Change: tl_settings.fernschach_resetRecords.datum -> date statt datim, weil nur das Datum benötigt wird
-* Add: Model HauptkontoModel für den Zugriff auf tl_fernschach_spieler_konto´
+* Add: Model HauptkontoModel für den Zugriff auf tl_fernschach_spieler_konto
 ACHTUNG: Das Model funktioniert nicht
 
 ## Version 0.20.1 (2024-07-04)